Add Project to ProjectContext during directory parsing
A line has been added in the ProjectUtil java file to add the current project to the
ProjectContext's projectList when the codeProject is set as a git directory.
This change ensures that all Git projects are added to the list when the directory
structure is parsed, allowing for easier tracking and management of Git projects within the application.